// src/types/database.types.ts
export type Json =
  | string
  | number
  | boolean
  | null
  | { [key: string]: Json | undefined }
  | Json[]

export type Database = {
  public: {
    Tables: {
      agents: {
        Row: {
          character_card: string | null
          color: string
          created_at: string
          creator_id: number
          display_name: string
          earnings: number | null
          endpoint: string
          eth_wallet_address: string | null
          id: number
          image_url: string
          last_health_check: string | null
          platform: string
          single_sentence_summary: string | null
          sol_wallet_address: string | null
          status: string | null
          type: string
          updated_at: string
          uuid: string | null
        }
        Insert: {
          character_card?: string | null
          color: string
          created_at?: string
          creator_id: number
          display_name: string
          earnings?: number | null
          endpoint: string
          eth_wallet_address?: string | null
          id?: number
          image_url: string
          last_health_check?: string | null
          platform: string
          single_sentence_summary?: string | null
          sol_wallet_address?: string | null
          status?: string | null
          type?: string
          updated_at?: string
          uuid?: string | null
        }
        Update: {
          character_card?: string | null
          color?: string
          created_at?: string
          creator_id?: number
          display_name?: string
          earnings?: number | null
          endpoint?: string
          eth_wallet_address?: string | null
          id?: number
          image_url?: string
          last_health_check?: string | null
          platform?: string
          single_sentence_summary?: string | null
          sol_wallet_address?: string | null
          status?: string | null
          type?: string
          updated_at?: string
          uuid?: string | null
        }
        Relationships: [
          {
            foreignKeyName: "agents_creator_id_fkey"
            columns: ["creator_id"]
            isOneToOne: false
            referencedRelation: "users"
            referencedColumns: ["id"]
          },
        ]
      }
// ...
      rooms: {
        Row: {
          active: boolean
          chain_family: string
          chain_id: number
          color: string | null
          contract_address: string | null
          created_at: string
          creator_id: number
          game_master_action_log: Json | null
          game_master_id: number | null
          id: number
          image_url: string | null
          name: string
          participants: number
          pvp_action_log: Json | null
          room_config: Json | null
          type_id: number
          updated_at: string
        }
        Insert: {
          active?: boolean
          chain_family: string
          chain_id: number
          color?: string | null
          contract_address?: string | null
          created_at?: string
          creator_id: number
          game_master_action_log?: Json | null
          game_master_id?: number | null
          id?: number
          image_url?: string | null
          name: string
          participants?: number
          pvp_action_log?: Json | null
          room_config?: Json | null
          type_id: number
          updated_at?: string
        }
        Update: {
          active?: boolean
          chain_family?: string
          chain_id?: number
          color?: string | null
          contract_address?: string | null
          created_at?: string
          creator_id?: number
          game_master_action_log?: Json | null
          game_master_id?: number | null
          id?: number
          image_url?: string | null
          name?: string
          participants?: number
          pvp_action_log?: Json | null
          room_config?: Json | null
          type_id?: number
          updated_at?: string
        }
        Relationships: [
          {
            foreignKeyName: "rooms_creator_id_fkey"
            columns: ["creator_id"]
            isOneToOne: false
            referencedRelation: "users"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "rooms_game_master_id_fkey"
            columns: ["game_master_id"]
            isOneToOne: false
            referencedRelation: "agents"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "rooms_type_id_fkey"
            columns: ["type_id"]
            isOneToOne: false
            referencedRelation: "room_types"
            referencedColumns: ["id"]
          },
        ]
      }
// ...
      round_observations: {
        Row: {
          content: Json
          created_at: string
          creator: string | null
          id: number
          observation_type: string
          round_id: number
        }
        Insert: {
          content: Json
          created_at?: string
          creator?: string | null
          id?: number
          observation_type: string
          round_id: number
        }
        Update: {
          content?: Json
          created_at?: string
          creator?: string | null
          id?: number
          observation_type?: string
          round_id?: number
        }
        Relationships: [
          {
            foreignKeyName: "round_observations_round_id_fkey"
            columns: ["round_id"]
            isOneToOne: false
            referencedRelation: "rounds"
            referencedColumns: ["id"]
          },
        ]
      }
// ...
      rounds: {
        Row: {
          active: boolean
          created_at: string
          game_master_action_log: Json | null
          game_master_id: number | null
          id: number
          pvp_action_log: Json | null
          pvp_status_effects: Json | null
          room_id: number
          round_config: Json | null
          status: Database["public"]["Enums"]["round_status"]
          updated_at: string
        }
        Insert: {
          active: boolean
          created_at?: string
          game_master_action_log?: Json | null
          game_master_id?: number | null
          id?: number
          pvp_action_log?: Json | null
          pvp_status_effects?: Json | null
          room_id: number
          round_config?: Json | null
          status?: Database["public"]["Enums"]["round_status"]
          updated_at?: string
        }
        Update: {
          active?: boolean
          created_at?: string
          game_master_action_log?: Json | null
          game_master_id?: number | null
          id?: number
          pvp_action_log?: Json | null
          pvp_status_effects?: Json | null
          room_id?: number
          round_config?: Json | null
          status?: Database["public"]["Enums"]["round_status"]
          updated_at?: string
        }
        Relationships: [
          {
            foreignKeyName: "rounds_room_id_fkey"
            columns: ["room_id"]
            isOneToOne: false
            referencedRelation: "rooms"
            referencedColumns: ["id"]
          },
        ]
      }
// ...
    }
// ...
  }
}

type PublicSchema = Database[Extract<keyof Database, "public">]

export type Tables<
  PublicTableNameOrOptions extends
    | keyof (PublicSchema["Tables"] & PublicSchema["Views"])
    | { schema: keyof Database },
  TableName extends PublicTableNameOrOptions extends { schema: keyof Database }
    ? keyof (Database[PublicTableNameOrOptions["schema"]]["Tables"] &
        Database[PublicTableNameOrOptions["schema"]]["Views"])
    : never = never,
> = PublicTableNameOrOptions extends { schema: keyof Database }
  ? (Database[PublicTableNameOrOptions["schema"]]["Tables"] &
      Database[PublicTableNameOrOptions["schema"]]["Views"])[TableName] extends {
      Row: infer R
    }
    ? R
    : never
  : PublicTableNameOrOptions extends keyof (PublicSchema["Tables"] &
        PublicSchema["Views"])
    ? (PublicSchema["Tables"] &
        PublicSchema["Views"])[PublicTableNameOrOptions] extends {
        Row: infer R
      }
      ? R
      : never
    : never
// ...
export type Enums<
  PublicEnumNameOrOptions extends
    | keyof PublicSchema["Enums"]
    | { schema: keyof Database },
  EnumName extends PublicEnumNameOrOptions extends { schema: keyof Database }
    ? keyof Database[PublicEnumNameOrOptions["schema"]]["Enums"]
    : never = never,
> = PublicEnumNameOrOptions extends { schema: keyof Database }
  ? Database[PublicEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : PublicEnumNameOrOptions extends keyof PublicSchema["Enums"]
    ? PublicSchema["Enums"][PublicEnumNameOrOptions]
    : never